bot/requirements.txt
undetected-chromedriver
setuptools
livekit
livekit-api
numpy
python-dotenv
pyyaml
grpcio
